Update to permit `user_id`
In order for students to pass in a `user_id` whenever we make a `POST http://localhost:3000/api/v1/notes`, we need to add this into the permitted params. 

So whenever a student passes in a JSON of `{ "title": "test-title", "body": "test-body", "user_id": "YOUR_USER_ID"}`, the association can be made. Since the `models/notes.rb` has an optional parameter, it is okay to be omitted for the rest of the example to work. However, if they want to implement the user association, the need to pass in the ID since we are using strong parameters.